# apps/rust/openapi_output.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Freefire",
    "description": "api gratis",
    "license": {
      "name": ""
    },
    "version": "0.0.1"
  },
  "paths": {
    "/api/anime": {
      "get": {
        "tags": [
          "anime"
        ],
        "operationId": "anime_index",
        "responses": {
          "200": {
            "description": "Handles GET requests for the anime endpoint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimeResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/anime/search": {
      "get": {
        "tags": [
          "anime"
        ],
        "operationId": "anime_search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search parameter for filtering results",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "sample_value"
          }
        ],
        "responses": {
          "200": {
            "description": "Searches for anime based on query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/anime2": {
      "get": {
        "tags": [
          "anime2"
        ],
        "operationId": "anime2_index",
        "responses": {
          "200": {
            "description": "Handles GET requests for the anime2 endpoint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Anime2Response"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/anime2/search": {
      "get": {
        "tags": [
          "anime2"
        ],
        "operationId": "anime2_search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search parameter for filtering results",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "sample_value"
          }
        ],
        "responses": {
          "200": {
            "description": "Searches for anime2 based on query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/komik/search": {
      "get": {
        "tags": [
          "komik"
        ],
        "operationId": "komik_search",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "description": "Search parameter for filtering results",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "sample_value"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination (starts from 1)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Searches for komik based on query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Anime2Data": {
        "type": "object",
        "required": [
          "ongoing_anime",
          "complete_anime"
        ],
        "properties": {
          "complete_anime": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompleteAnimeItem"
            }
          },
          "ongoing_anime": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OngoingAnimeItem"
            }
          }
        }
      },
      "Anime2Response": {
        "type": "object",
        "required": [
          "status",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Anime2Data"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "AnimeData": {
        "type": "object",
        "required": [
          "ongoing_anime",
          "complete_anime"
        ],
        "properties": {
          "complete_anime": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompleteAnimeItem"
            }
          },
          "ongoing_anime": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OngoingAnimeItem"
            }
          }
        }
      },
...
      "AnimeResponse": {
        "type": "object",
        "required": [
          "status",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AnimeData"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "CompleteAnimeItem": {
        "type": "object",
        "required": [
          "title",
          "slug",
          "poster",
          "episode_count",
          "anime_url"
        ],
        "properties": {
          "anime_url": {
            "type": "string"
          },
          "episode_count": {
            "type": "string"
          },
          "poster": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "MangaItem": {
        "type": "object",
        "required": [
          "title",
          "poster",
          "chapter",
          "score",
          "date",
          "type",
          "slug"
        ],
        "properties": {
          "chapter": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "poster": {
            "type": "string"
          },
          "score": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "OngoingAnimeItem": {
        "type": "object",
        "required": [
          "title",
          "slug",
          "poster",
          "score",
          "anime_url"
        ],
        "properties": {
          "anime_url": {
            "type": "string"
          },
          "poster": {
            "type": "string"
          },
          "score": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "required": [
          "current_page",
          "last_visible_page",
          "has_next_page",
          "has_previous_page"
        ],
        "properties": {
          "current_page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "has_next_page": {
            "type": "boolean"
          },
          "has_previous_page": {
            "type": "boolean"
          },
          "last_visible_page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "next_page": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "previous_page": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "SearchResponse": {
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MangaItem"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      }
    }
  },
...
}